Single Channel Puncta Analysis
Based on the original dendrite analysis macro, this performs puncta analysis on one channel but uses individual thresholding for the analysis.
A .csv file is required for this macro. Ensure it includes column headers "File" for file names and "Threshold" for thresholding values.

Similarly, there is an option to include ROI length for calculation of puncta/length. .csv column titles should be "File" and "Length". Can also be adjusted for area.